requirements.txt
PyGObject~=3.52.3
pycairo~=1.27.0
PyOpenGL==3.1.9